Add option not to delete target directory on failure
When a metastore operation fails, especially due to timeout we don't
know whether it ultimately completes. Cleaning up new target
directories can cause data loss under the following set of
circumstances.

1. a query reads from and writes to the same partition
2. the metastore alter table call times out, but eventually completes
3. the metastore calls to undo the alter table don't succeed

If the query does not read from and write to the same partition, it
won't cause data loss but there will still be an invalid entry in the
metastore.

Therefore, we add an option not to cleanup the target directory on
failure when the write mode is DIRECT_TO_TARGET_NEW_DIRECTORY, i.e.
where we write the new partition directly to a new target directory.