Delay stage creation until ready for execution
This is in preparation for stage retries.  we don't want any of the
stage info to be generated until it's ready for execution so we can be
certain that no existing stages need to be updated (e.g. we would need
to update the schedulers for later stages with new output information if
a child stage failed)