Make columnEncodings into a map
When dwrf encryption support is added, we can't rely on list position
of the encoding to be the same as the column index because encodings
for encrypted columns will be stored separately in the StripeEncryptionGroups,
and the reader may not even have access to them if the user doesn't have
the key for those groups. Instead we explicitly map the column index to the
correct encoding.